+++ AverageJoes.Models/User/UserCreate.cs
@@ -20,7 +20,7 @@ namespace AverageJoes.Models.User
         [Required]
         public string Email { get; set; }
         [Required]
-        public int CreditCard { get; set; }
+        public long CreditCard { get; set; }
         [ForeignKey(nameof(Memberships))]
         public int MembershipID { get; set; }
         public virtual Memberships Membership { get; set; }
